--- a/src/FlowtideDotNet.Core/ColumnStore/Utils/AvxUtils.cs
+++ b/src/FlowtideDotNet.Core/ColumnStore/Utils/AvxUtils.cs
@@ -210,7 +210,7 @@ namespace FlowtideDotNet.Core.ColumnStore.Utils
             }
         }
 
-        public static unsafe void InPlaceMemCopyAdditionByType(Span<int> array, Span<sbyte> typeIds, int sourceIndex, int destIndex, int length, Span<int> toAdd)
+        public static unsafe void InPlaceMemCopyAdditionByTypeScalar(Span<int> array, Span<sbyte> typeIds, int sourceIndex, int destIndex, int length, Span<int> toAdd)
         {
             unsafe
             {
@@ -235,6 +235,84 @@ namespace FlowtideDotNet.Core.ColumnStore.Utils
                             var typeId = typeIds[sourceIndex + i];
                             array[destIndex + i] = array[sourceIndex + i] + toAdd[typeId];
                         }
+                    }
+                }
+            }
+        }
+
+        public static unsafe void InPlaceMemCopyAdditionByType(
+            Span<int> array, 
+            Span<sbyte> typeIds, 
+            int sourceIndex, 
+            int destIndex, 
+            int length, 
+            Span<int> toAdd,
+            int numberOfTypes)
+        {
+            if (!Avx2.IsSupported || numberOfTypes > 8)
+            {
+                // Fallback to scalar implementation if AVX2 is not supported
+                // or if the number of types is greater than 8.
+                InPlaceMemCopyAdditionByTypeScalar(array, typeIds, sourceIndex, destIndex, length, toAdd);
+                return;
+            }
+            // Avx implementation if the number of types are less than 8.
+            // This allows using PermuteVar8x32 to get the correct additions.
+            fixed (int* pArray = array)
+            fixed (sbyte* pTypeIds = typeIds)
+            fixed (int* pToAdd = toAdd)
+            {
+                int vecLength = Vector256<int>.Count;
+                Vector256<int> valueToAddVec = Vector256.Load(pToAdd);
+
+                if (sourceIndex < destIndex && sourceIndex + length > destIndex)
+                {
+                    // Process from end to start (backward) to handle overlap
+                    for (int j = length - vecLength; j >= 0; j -= vecLength)
+                    {
+                        // Load 8 ints from source array and 8 typeids
+                        Vector256<int> srcVec = Avx2.LoadVector256(pArray + sourceIndex + j);
+                        Vector128<sbyte> typeIdVec = Avx2.LoadVector128(pTypeIds + sourceIndex + j);
+
+                        var typeIdVecInt = Avx2.ConvertToVector256Int32(typeIdVec);
+                        var additions = Avx2.PermuteVar8x32(valueToAddVec, typeIdVecInt);
+
+                        // Mask the source vector to conditionally add valueToAdd
+                        Vector256<int> addedVec = Avx2.Add(srcVec, additions);
+
+                        // Store the result
+                        Avx2.Store(pArray + destIndex + j, addedVec);
+                    }
+
+                    for (int j = (length % vecLength) - 1; j >= 0; j--)
+                    {
+                        var typeId = pTypeIds[sourceIndex + j];
+                        pArray[destIndex + j] = pArray[sourceIndex + j] + pToAdd[typeId];
+                    }
+                }
+                else
+                {
+                    int i;
+                    for (i = 0; i <= length - vecLength; i += vecLength)
+                    {
+                        // Load 8 ints from source array and 8 typeids
+                        Vector256<int> srcVec = Avx2.LoadVector256(pArray + sourceIndex + i);
+                        Vector128<sbyte> typeIdVec = Avx2.LoadVector128(pTypeIds + sourceIndex + i);
+
+                        var typeIdVecInt = Avx2.ConvertToVector256Int32(typeIdVec);
+                        var additions = Avx2.PermuteVar8x32(valueToAddVec, typeIdVecInt);
+
+                        // Use the mask to conditionally add valueToAddVec
+                        Vector256<int> addedVec = Avx2.Add(srcVec, additions);
+
+                        // Store the result
+                        Avx2.Store(pArray + destIndex + i, addedVec);
+                    }
+
+                    for (; i < length; i++)
+                    {
+                        var typeId = typeIds[sourceIndex + i];
+                        array[destIndex + i] = array[sourceIndex + i] + toAdd[typeId];
                     }
                 }
             }