Also fix some performance issues in map and union column

Benchmarks:

Row based:

| Method                    | Mean    | Error    | StdDev   | Processed Events / s |
|-------------------------- |--------:|---------:|---------:|---------------------:|
| ListAggWithMapAggregation | 3.553 s | 0.0709 s | 0.0315 s |               316218 |

Before change:

| Method                    | Mean    | Error    | StdDev   | Processed Events / s |
|-------------------------- |--------:|---------:|---------:|---------------------:|
| ListAggWithMapAggregation | 1.688 s | 0.0973 s | 0.0579 s |               672820 |


After change:

| Method                    | Mean    | Error    | StdDev   | Processed Events / s |
|-------------------------- |--------:|---------:|---------:|---------------------:|
| ListAggWithMapAggregation | 1.463 s | 0.0374 s | 0.0247 s |               775520 |